chore(commands): create fetch, convert directory
